+++ src/main.c
@@ -1,4 +1,8 @@
+#define _POSIX_C_SOURCE 200809L
+
+#include <errno.h>
 #include <stddef.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
@@ -16,15 +20,16 @@ static Plugin* plugins[] = {
 };
 
 
-static void cleanup(Array resources, Error error) {
+static void cleanup(Array resources, Error* error) {
     if (error) {
-        fprintf(stderr, "%s\n", error);
+        fprintf(stderr, "%s\n", *error);
     }
 
     if (resources != NULL) {
         for (size_t i = 0; i < resources->length; ++i) {
             Resource_delete((Resource) resources->data[i]);
         }
+
         Array_delete(resources);
     }
 
@@ -37,51 +42,74 @@ static void cleanup(Array resources, Error error) {
 }
 
 
+static void flush_input(int input_fd, int output_fd, Error* error) {
+    ssize_t bytes_read;
+    const int BUFFER_SIZE = 4 * 1024;
+    uint8_t buffer[BUFFER_SIZE];
+
+    while ((bytes_read = read(input_fd, buffer, BUFFER_SIZE)) > 0) {
+        uint8_t* next_write = buffer;
+
+        while (bytes_read > 0) {
+            ssize_t bytes_written = write(
+                output_fd, next_write, (size_t) bytes_read);
+
+            if (bytes_written == -1) {
+                Error_errno(error, errno);
+                return;
+            }
+
+            bytes_read -= bytes_written;
+            next_write += bytes_written;
+        }
+    }
+
+    Error_clear(error);
+}
+
+
 int main(int argc, char* argv[]) {
     Error error;
-    Array resources = Options_parse(
+    int output_fd = STDOUT_FILENO;
+
+    Array inputs = Options_parse(
         argc, argv, plugins, STATIC_ARRAY_LENGTH(plugins), &error);
 
-    if (resources == NULL) {
-        cleanup(NULL, error ? error : NULL);
+    if (inputs == NULL) {
+        cleanup(NULL, error ? &error : NULL);
         return EXIT_SUCCESS;
     }
 
     for (size_t i = 0; i < STATIC_ARRAY_LENGTH(plugins); ++i) {
         if (plugins[i] != NULL) {
-            plugins[i]->run(resources, plugins[i]->options, &error);
+            plugins[i]->run(inputs, plugins[i]->options, &output_fd, &error);
 
             if (error) {
                 fprintf(stderr, "%s: %s\n", plugins[i]->get_name(), error);
-                cleanup(resources, NULL);
+                cleanup(inputs, NULL);
                 return EXIT_FAILURE;
             }
         }
     }
 
-    for (size_t i = 0; i < resources->length; ++i) {
-        Resource resource = (Resource) resources->data[i];
+    for (size_t i = 0; i < inputs->length; ++i) {
+        Resource input = (Resource) inputs->data[i];
 
-        if (resource == NULL) {
-            continue;
-        }
+        if (input != NULL) {
+            int input_fd = input->fd;
 
-        int fd = resource->fd;
+            if (input_fd != RESOURCE_NO_FD) {
+                flush_input(input_fd, output_fd, &error);
+                close(input_fd);
 
-        if (fd != RESOURCE_NO_FD) {
-            ssize_t nr_bytes_read;
-            const int BUFFER_SIZE = 256;
-            char buffer[BUFFER_SIZE];
-
-            while ((nr_bytes_read = read(fd, buffer, BUFFER_SIZE - 1)) > 0) {
-                buffer[nr_bytes_read] = '\0';
-                fputs(buffer, stdout);
+                if (Error_has(&error)) {
+                    cleanup(inputs, &error);
+                    return EXIT_FAILURE;
+                }
             }
-
-            close(fd);
         }
     }
 
-    cleanup(resources, NULL);
+    cleanup(inputs, NULL);
     return EXIT_SUCCESS;
 }